[feat]: 书签分类和书签列表增加 user_id 字段

+++ src/app/Model/BookClass.php
@@ -15,6 +15,7 @@ class BookClass extends NotORM {
     const KEY_NAME = 'name';
     const KEY_MODIFY_AT = 'modify_at';
     const KEY_CREATE_AT = 'create_at';
+    const KEY_USER_ID = 'user_id';
 
     public function getTableName($id)
     {
@@ -22,13 +23,14 @@ class BookClass extends NotORM {
     }
 
     /**
+     * @param int $user_id 用户id
      * @param $parent_id
      * @param $name
      * @return string
      */
-    public function addClass($parent_id, $name)
+    public function addClass($user_id, $parent_id, $name)
     {
-        $data = array(self::KEY_PARENT_ID => $parent_id, self::KEY_NAME => $name, self::KEY_CREATE_AT => time());
+        $data = array(self::KEY_USER_ID => $user_id, self::KEY_PARENT_ID => $parent_id, self::KEY_NAME => $name, self::KEY_CREATE_AT => time());
 
         $orm = $this->getORM();
         $orm->insert($data);
@@ -40,34 +42,37 @@ class BookClass extends NotORM {
      * 删除指定分类.
      *
      * @param int $id 分类id.
+     * @param int $user_id 用户id
      * @return bool|int
      * @throws \Exception
      */
-    public function deleteClass($id)
+    public function deleteClass($id, $user_id)
     {
-        return $this->getORM()->where(self::KEY_ID, $id)->delete();
+        return $this->getORM()->where(self::KEY_ID, $id)->where(self::KEY_USER_ID, $user_id)->delete();
     }
 
     /**
      * 获取某个分类下子分类的数量.
      *
      * @param int $id 指定 id
+     * @param int $user_id 用户id
      * @return int
      */
-    public function getCount($id)
+    public function getCount($id, $user_id)
     {
-        return $this->getORM()->where(self::KEY_PARENT_ID, $id)->count();
+        return $this->getORM()->where(self::KEY_PARENT_ID, $id)->where(self::KEY_USER_ID, $user_id)->count();
     }
 
     /**
      * 更新指定分类.
      *
+     * @param int $user_id 用户id.
      * @param int $id 分类id.
      * @param int $parent_id 父类id
      * @param string $name 分类名称
      * @return false|int
      */
-    public function updateClass($id, $parent_id, $name)
+    public function updateClass($user_id, $id, $parent_id, $name)
     {
         $orm = $this->getORM();
         $data = array(
@@ -76,7 +81,7 @@ class BookClass extends NotORM {
             self::KEY_MODIFY_AT => time()
         );
 
-        return $orm->where(self::KEY_ID, $id)->update($data);
+        return $orm->where(self::KEY_ID, $id)->where(self::KEY_USER_ID, $user_id)->update($data);
     }
 
     /**
@@ -85,10 +90,10 @@ class BookClass extends NotORM {
      * @param $id
      * @return bool
      */
-    public function exists($id)
+    public function exists($user_id, $id)
     {
         $orm = $this->getORM();
-        $count = $orm->where(self::KEY_ID, $id)->count();
+        $count = $orm->where(self::KEY_ID, $id)->where(self::KEY_USER_ID, $user_id)->count();
 
         return $count > 0;
     }
