+++ src/app/Model/BookMark.php
@@ -17,6 +17,7 @@ class BookMark extends NotORM
     const KEY_CLASS_ID = 'class_id';
     const KEY_CREATE_AT = 'create_at';
     const KEY_MODIFY_AT = 'modify_at';
+    const KEY_USER_ID = 'user_id';
 
     public function getTableName($id)
     {
@@ -27,38 +28,42 @@ class BookMark extends NotORM
      * 获取某个分类下书签的数量.
      *
      * @param int $class_id 分类id
+     * @param int $user_id 用户id
      * @return int
      */
-    public function getCount($class_id)
+    public function getCount($class_id, $user_id)
     {
-        return $this->getORM()->where(self::KEY_CLASS_ID, $class_id)->count();
+        return $this->getORM()->where(self::KEY_CLASS_ID, $class_id)->where(self::KEY_USER_ID, $user_id)->count();
     }
 
     /**
      * 获取所有的书签列表.
      *
+     * @param int $user_id 用户id
      * @return mixed
      */
-    public function getAllBookMark()
+    public function getAllBookMark($user_id)
     {
-        return $this->getORM()->fetchAll();
+        return $this->getORM()->where(self::KEY_USER_ID, $user_id)->fetchAll();
     }
 
     /**
      * 添加新的书签.
      *
+     * @param int $user_id 用户id
      * @param int $class_id 分类id
      * @param string $mark_name 书签名称
      * @param string $url url地址
      * @return string
      */
-    public function addBookMark($class_id, $mark_name, $url)
+    public function addBookMark($user_id, $class_id, $mark_name, $url)
     {
         $orm = $this->getORM();
         $data = array(
             self::KEY_CLASS_ID => $class_id,
             self::KEY_MARK_NAME => $mark_name,
             self::KEY_URL => $url,
+            self::KEY_USER_ID => $user_id
         );
 
         $orm->insert($data);
@@ -70,12 +75,13 @@ class BookMark extends NotORM
      * 判断 url 是否存在.
      *
      * @param string $url 要判断的 url
+     * @param int $user_id 用户id.
      * @return bool
      */
-    public function urlExists($url)
+    public function urlExists($url, $user_id)
     {
         $orm = $this->getORM();
-        $count = $orm->where(self::KEY_URL, $url)->count();
+        $count = $orm->where(self::KEY_URL, $url)->where(self::KEY_USER_ID, $user_id)->count();
 
         return $count > 0;
     }
@@ -84,25 +90,27 @@ class BookMark extends NotORM
      * 删除指定的书签.
      *
      * @param int $id 要删除的id.
+     * @param int $user_id 用户id.
      * @return bool|int
      * @throws \Exception
      */
-    public function deleteBookMark($id)
+    public function deleteBookMark($id, $user_id)
     {
         $orm = $this->getORM();
-        return $orm->where(self::KEY_ID, $id)->delete();
+        return $orm->where(self::KEY_ID, $id)->where(self::KEY_USER_ID, $user_id)->delete();
     }
 
     /**
      * 修改指定书签.
      *
+     * @param int $user_id 用户id
      * @param int $id 要修改的书签
      * @param string $name 书签
      * @param string $url 书签地址
      * @param $class_id 分类id
      * @return int
      */
-    public function updateBookMark($id, $name, $url, $class_id)
+    public function updateBookMark($user_id, $id, $name, $url, $class_id)
     {
         $orm = $this->getORM();
         $data = array(
@@ -112,19 +120,20 @@ class BookMark extends NotORM
             self::KEY_MODIFY_AT => time()
         );
 
-        return $orm->where(self::KEY_ID, $id)->update($data);
+        return $orm->where(self::KEY_ID, $id)->where(self::KEY_USER_ID, $user_id)->update($data);
     }
 
 
     /**
      * 获取分类下的所有书签.
      *
+     * @param int $user_id 用户id
      * @param int $class_id 要获取的分类id
      * @return mixed
      */
-    public function getBookMarkFromClass($class_id)
+    public function getBookMarkFromClass($user_id, $class_id)
     {
         $orm = $this->getORM();
-        return $orm->where(self::KEY_CLASS_ID, $class_id)->fetchAll();
+        return $orm->where(self::KEY_CLASS_ID, $class_id)->where(self::KEY_USER_ID, $user_id)->fetchAll();
     }
 }
